Add Wasmtime Client
