+++ Cargo.toml
@@ -23,4 +23,5 @@ reqwest = { version = "0.11.10", features = ["blocking", "json", "native-tls"] }
 derive-new = "0.5.9"
 http = "0.2.8"
 base64 = "0.13.0"
-
+wasi-experimental-http = "0.10.0"
+bytes = "1"
