+++ src/client/mod.rs
@@ -1,5 +1,8 @@
 mod common;
 mod openwhisk_client;
-
+mod native_client;
+mod wasmtime_client;
 pub use common::*;
 pub use openwhisk_client::OpenwhiskClient;
+pub use native_client::NativeClient;
+pub use wasmtime_client::WasmClient;
